pyproject.toml
[tool.poetry]
name = "dropbox_clone"
version = "0.1.0"
description = "a project the clones dropbox with some modifications to test how dropbox works. It is also a proof of concept project."
authors = ["Mustafa Prodhan <mustafa.prodhan.317@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
django-environ = "^0.4.5"
autopep8 = "^1.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
